+++ src/main/java/just/learn/service/UserService.java
@@ -9,11 +9,11 @@ public interface UserService {
 
     public User insert(User user);
 
-    public boolean delete(Long id);
+    public boolean delete(Integer id);
 
     public int update(User user);
 
-    public User getById(Long id);
+    public User getById(Integer id);
 
     PageQueryBean getLimitObjects(QueryCondition condition);
 //public   List<User> getAll();
